requirements.txt
# Django & Web Framework
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-decouple==3.8

# Production Server & Database
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==1.3.0
whitenoise==6.6.0
dj-static==0.0.6

# Security & Authentication
django-environ==0.10.0
cryptography==41.0.7

# Desktop Application Dependencies
pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.13
vosk==0.3.45
google-generativeai==0.3.0
pyautogui==0.9.53
python-dotenv==1.0.0
requests==2.31.0

# Caching & Performance
redis==5.0.0
django-redis==5.4.0

# Utilities
pillow==10.1.0
celery==5.3.4
click==8.1.7

# Monitoring & Logging
sentry-sdk==1.39.1
django-extensions==3.2.3

# Development Tools (comment out in production)
# django-debug-toolbar==4.2.0
# pytest==7.4.3
# pytest-django==4.7.0
# black==23.12.1
# flake8==6.1.0

